Converts project to a gem. Fixes #4.
