+++ .gitignore
@@ -1,13 +1,25 @@
+/.bundle/
+/.yardoc
+/Gemfile.lock
+/_yardoc/
+/coverage/
+/doc/
+/pkg/
+/spec/reports/
+/tmp/
+*.bundle
+*.so
+*.o
+*.a
 *~
 *_test.*
 rotation.txt
 test*.txt
 results.txt
-.bundle/*
-bin/*
 rotate.txt
 \#*\#
 .\#*
 tmp
 vendor
 *.swp
+mkmf.log
