+++ Rakefile
@@ -1,0 +1,1 @@
+require "bundler/gem_tasks"
