+++ lib/mob_rotation.rb
@@ -1,0 +1,7 @@
+require "mob_rotation/version"
+require "mob_rotation/database"
+require "mob_rotation/rotation"
+
+# Manage a mob list during a mobbing session
+module MobRotation
+end
