+++ lib/mob_rotation/mobster.rb
@@ -1,0 +1,12 @@
+module MobRotation
+  class Mobster
+    attr_reader :name, :email
+
+    def initialize(name, email = nil)
+      @name = name
+      @email = email
+    end
+
+    alias_method :to_s, :name
+  end
+end
