+++ lib/mob_rotation/rotation.rb
@@ -1,0 +1,210 @@
+require "time"
+require "mob_rotation/command_routing"
+require "mob_rotation/mobster"
+
+module MobRotation
+  class Rotation
+    include CommandRouting
+    extend CommandRouting::ClassMethods
+
+    define_command("show") do
+      show_mobsters
+    end
+
+    define_command("run_with_timer") do
+      countdown_to_rotate(ARGV[1].to_i)
+    end
+
+    define_command("rotate") do
+      rotate
+      show_mobsters
+    end
+
+    define_command("random") do
+      random(ARGV[1])
+      show_mobsters
+    end
+
+    define_command("add") do |*mobster_names|
+      add_mobster(*mobster_names)
+      show_mobsters
+    end
+
+    define_command("remove") do |*mobster_names|
+      remove_mobsters(*mobster_names)
+      show_mobsters
+    end
+
+    define_command("help") do
+      show_help
+    end
+
+    def self.annoying_and_probably_not_accidental_beep(n = number_of_beeps)
+      n.times { print("\a") || sleep(minimum_sleep_between_beeps) }
+    end
+
+    def self.minimum_sleep_between_beeps
+      0.2
+    end
+
+    def self.number_of_beeps
+      5
+    end
+
+    def initialize(database, git_dir)
+      @git_dir = git_dir
+      @database = database
+      @real_mobsters = @database.sanitized_entries_in do | entry |
+        build_mobster(entry)
+      end
+    end
+
+    def build_mobster(entry)
+      name = extract_name_from(entry)
+      email = extract_email_from(entry)
+      Mobster.new(name, email)
+    end
+
+    def extract_email_from(entry)
+      if entry =~ /\<(.*)\>/
+        Regexp.last_match[1]
+      end
+    end
+
+    def write(text)
+      puts(text)
+    end
+
+    def show_mobsters
+      @real_mobsters.each_with_index do |person, index|
+        case index
+        when 0
+          write("git username: #{person.name}")
+          write("git user email: #{person.email}")
+          format_mobster("Driver", person)
+        when 1
+          format_mobster("Navigator", person)
+        else
+          format_mobster("Mobster", person)
+        end
+      end
+    end
+
+    def format_mobster(role, person)
+      if person.email.to_s.empty?
+        write("#{role} #{person.name}")
+      else
+        write("#{role} #{person.name.strip} <#{person.email.strip}>")
+      end
+    end
+
+    def add_mobster(*mobsters_to_add)
+      mobsters_to_add.collect(&:to_s).collect(&:strip).each do |mobster_to_add|
+        fail if mobster_to_add.empty?
+
+        if @real_mobsters.collect(&:name).include?(mobster_to_add)
+          write("user name '#{mobster_to_add}' already exists")
+          next
+        end
+
+        @real_mobsters << Mobster.new(mobster_to_add)
+      end
+
+      sync
+    end
+
+    def remove_mobster(given_mobster)
+      @real_mobsters.each_with_index do |mobster, i|
+        if found_mobster(mobster, given_mobster)
+          @real_mobsters.delete_at(i)
+        end
+      end
+      sync
+    end
+
+    def remove_mobsters(*mobsters)
+      mobsters.each do |mobster|
+        remove_mobster(mobster)
+      end
+    end
+
+    def show_help
+      puts ["Available commands are:",
+            "show",
+            "help",
+            "rotate",
+            "random",
+            "add <name1> [name2]",
+            "remove <name1> [name2]",
+            "run_with_timer [seconds]"
+           ]
+    end
+
+
+    def countdown_to_rotate(seconds = 300)
+      sleep(seconds)
+      puts("Time to rotate")
+      self.class.annoying_and_probably_not_accidental_beep
+    end
+
+    def inform_lovely_user(command)
+      puts("Unknown command #{command}")
+      show_help
+    end
+
+    def rotate_mobsters(&rotation_algorithm)
+      rotation_algorithm.call
+      git_config_update
+      sync
+    end
+
+    def rotate
+      puts("")
+      rotate_mobsters do
+        @real_mobsters << @real_mobsters.shift
+      end
+    end
+
+    def random(seed = nil)
+      puts("Randomized Output")
+      rotate_mobsters do
+        srand(seed.to_i) if seed
+        @real_mobsters.shuffle!
+      end
+    end
+
+    def extract_next_mobster_email
+      email = @real_mobsters.first.email.to_s.strip
+      email.empty? ? "mob@rubysteps.com" : email
+    end
+
+    def time_to_rotate
+      puts("Time to rotate!")
+    end
+
+    private
+
+    def git_config_update
+      # FIX yo that rescue nil is bogus
+      system(git_config_for("name", @real_mobsters.first.name)) rescue nil
+      system(git_config_for("email", extract_next_mobster_email)) rescue nil
+    end
+
+    def git_config_for(property, value)
+      "git --git-dir=#{@git_dir} config user.#{property} '#{value}'"
+    end
+
+    def sync
+      @database.write(@real_mobsters)
+    end
+
+    def found_mobster(line, mobster)
+      line.to_s.strip == mobster.to_s
+    end
+
+    def extract_name_from(entry)
+      entry_to_array = entry.split("")
+      entry_to_array.take_while { |c| c != "<" }.join("")
+    end
+  end
+end
