+++ lib/mob_rotation/version.rb
@@ -1,0 +1,3 @@
+module MobRotation
+  VERSION = "0.1.0"
+end
