--- mob_rotation
@@ -1,10 +1,0 @@
-#!/usr/bin/env ruby
-require_relative 'mob_rotation'
-
-file_name = ENV['DB_FILE'] || 'rotate.txt'
-command = ARGV[0] || 'show'
-mobster_names = ARGV[1..-1]
-git_dir = ENV['MOB_GIT_DIR'] || File.join(Dir.pwd, ".git")
-
-mob_rotator = MobRotation.new(Database.new(file_name), git_dir)
-mob_rotator.command_router(command, mobster_names)
