--- mobster.rb
@@ -1,11 +1,0 @@
-class Mobster
-  attr_reader :name, :email
-  def initialize(name, email=nil)
-    @name = name
-    @email = email
-  end
-
-  def to_s
-    @name
-  end
-end
