+++ spec/rotation_spec.rb
@@ -1,14 +1,18 @@
-require_relative 'spec_helper'
+require "spec_helper"
 
-describe MobRotation do
+describe MobRotation::Rotation do
   extend FooFighter
 
-  let(:mob_rotator) { MobRotation.new(Database.new(file_name), "./tmp/test_project/.git") }
+  let(:mob_rotator) {
+    MobRotation::Rotation.new(
+      MobRotation::Database.new(file_name),
+      "./tmp/test_project/.git"
+    )
+  }
 
   before { FileUtils.rm_f(file_name) }
 
   describe "#show_mobsters" do
-
     let(:file_name) { "test.txt" }
     let(:output) { Output.new }
 
@@ -19,7 +23,7 @@ describe MobRotation do
 
       $stdout = output
 
-      mob_rotator.show_mobsters()
+      mob_rotator.show_mobsters
     end
 
     it "print out all the mobsters in the text file" do
@@ -39,36 +43,36 @@ describe MobRotation do
     let(:file_name) { "test.txt" }
 
     it "adds a mobster to the file" do
-      mob_rotator.add_mobster 'Jackie'
-      mob_rotator.add_mobster 'Phil'
+      mob_rotator.add_mobster("Jackie")
+      mob_rotator.add_mobster("Phil")
 
       expect(File.read(file_name)).to eq("Jackie\nPhil\n")
     end
 
     it "accepts more than one mobster" do
-      mob_rotator.add_mobster "Jackie", "Phil"
+      mob_rotator.add_mobster("Jackie", "Phil")
       expect(File.read(file_name)).to include("\nPhil")
     end
 
     it "skips adding duplicate mobster name" do
-      mob_rotator.add_mobster "Rosie", "Rosie"
+      mob_rotator.add_mobster("Rosie", "Rosie")
       expect(File.read(file_name)).to eq("Rosie\n")
     end
 
     it "strips leading and trailing whitespace" do
-      mob_rotator.add_mobster " \n   Rosie  \n  "
+      mob_rotator.add_mobster(" \n   Rosie  \n  ")
       expect(File.read(file_name)).to eq("Rosie\n")
     end
 
     it "does not allow blank usernames" do
-      expect { mob_rotator.add_mobster " " }.
-        to raise_error
+      expect { mob_rotator.add_mobster(" ") }
+        .to raise_error
     end
   end
-  
+
   describe "#remove_mobster" do
     let(:file_name) { "test_remove.txt" }
-    
+
     before do
       File.open(file_name, "w") do |file|
         file << "Bob\n" << "Phoebe\n" << "Joe"
@@ -76,25 +80,23 @@ describe MobRotation do
     end
 
     it "removes a mobster from the file" do
-      mob_rotator.remove_mobster "Bob"
-      expect(File.read(file_name)).not_to include('Bob')
-
+      mob_rotator.remove_mobster("Bob")
+      expect(File.read(file_name)).not_to include("Bob")
     end
   end
 
   describe "#rotate" do
     let(:file_name) { "test_remove.txt" }
-    
+
     before do
       File.open(file_name, "w") do |file|
         file << "Bob\n" << "Phoebe\n" << "Joe\n"
       end
     end
-    
+
     it "rotates mobsters" do
-      
       mob_rotator.rotate
-      # new file to equal 'Phoebe Joe Bob'
+      # new file to equal "Phoebe Joe Bob"
       expect(File.read(file_name)).to eq("Phoebe\nJoe\nBob\n")
     end
   end
@@ -105,16 +107,23 @@ describe MobRotation do
     after(:each) do
       FileUtils.rm_f(file_name)
     end
-    
+
     it "returns the email address" do
-      email = MobRotation.new(Database.new(file_name), "irrelevant git dir").extract_email_from('a <b@example.com>')
-      expect(email).to eq('b@example.com')
+      email = MobRotation::Rotation.new(
+        MobRotation::Database.new(file_name),
+        "irrelevant git dir"
+      ).extract_email_from("a <b@example.com>")
+
+      expect(email).to eq("b@example.com")
     end
 
     it "it handles arbitary email addresses" do
-      email = MobRotation.new(Database.new(file_name), "irrelevant git dir").extract_email_from('bob <bob@example.com>')
-      expect(email).to eq('bob@example.com')
-    end
+      email = MobRotation::Rotation.new(
+        MobRotation::Database.new(file_name),
+        "irrelevant git dir"
+      ).extract_email_from("bob <bob@example.com>")
 
+      expect(email).to eq("bob@example.com")
+    end
   end
 end
